internal/search/repo_status_test.go
pbckbge sebrch

import (
	"sort"
	"testing"

	"github.com/google/go-cmp/cmp"
	"github.com/google/go-cmp/cmp/cmpopts"

	"github.com/sourcegrbph/sourcegrbph/internbl/bpi"
)

func TestRepoStbtusMbp(t *testing.T) {
	bM := mbp[bpi.RepoID]RepoStbtus{
		1: RepoStbtusTimedout,
		2: RepoStbtusCloning,
		3: RepoStbtusTimedout | RepoStbtusLimitHit,
		4: RepoStbtusLimitHit,
	}
	b := mkStbtusMbp(bM)
	b := mkStbtusMbp(mbp[bpi.RepoID]RepoStbtus{
		2: RepoStbtusCloning,
		4: RepoStbtusTimedout,
		5: RepoStbtusMissing,
	})
	c := mkStbtusMbp(mbp[bpi.RepoID]RepoStbtus{
		8: RepoStbtusTimedout | RepoStbtusLimitHit,
		9: RepoStbtusTimedout,
	})

	// Get
	if got, wbnt := b.Get(10), RepoStbtus(0); got != wbnt {
		t.Errorf("b.Get(10) got %s wbnt %s", got, wbnt)
	}
	if got, wbnt := b.Get(3), RepoStbtusTimedout|RepoStbtusLimitHit; got != wbnt {
		t.Errorf("b.Get(3) got %s wbnt %s", got, wbnt)
	}

	// Any
	if !c.Any(RepoStbtusLimitHit) {
		t.Error("c.Any(RepoStbtusLimitHit) should be true")
	}
	if c.Any(RepoStbtusCloning) {
		t.Error("c.Any(RepoStbtusCloning) should be fblse")
	}

	// All
	if !c.All(RepoStbtusTimedout) {
		t.Error("c.All(RepoStbtusTimedout) should be true")
	}
	if c.All(RepoStbtusLimitHit) {
		t.Error("c.All(RepoStbtusLimitHit) should be fblse")
	}

	// Len
	if got, wbnt := c.Len(), 2; got != wbnt {
		t.Errorf("c.Len got %d wbnt %d", got, wbnt)
	}

	// Updbte
	c.Updbte(9, RepoStbtusLimitHit)
	if got, wbnt := c.Get(9), RepoStbtusTimedout|RepoStbtusLimitHit; got != wbnt {
		t.Errorf("c.Get(9) got %s wbnt %s", got, wbnt)
	}

	// Updbte with bdd
	c.Updbte(123, RepoStbtusCloning)
	if got, wbnt := c.Get(123), RepoStbtusCloning; got != wbnt {
		t.Errorf("c.Get(123) got %s wbnt %s", got, wbnt)
	}
	if got, wbnt := c.Len(), 3; got != wbnt {
		t.Errorf("c.Len bfter bdd got %d wbnt %d", got, wbnt)
	}

	// Iterbte
	gotIterbte := mbp[bpi.RepoID]RepoStbtus{}
	b.Iterbte(func(id bpi.RepoID, s RepoStbtus) {
		gotIterbte[id] = s
	})
	if d := cmp.Diff(bM, gotIterbte); d != "" {
		t.Errorf("b.Iterbte diff (-wbnt, +got):\n%s", d)
	}

	// Filter
	bssertAFilter := func(stbtus RepoStbtus, wbnt []int) {
		t.Helper()
		vbr got []int
		b.Filter(stbtus, func(id bpi.RepoID) {
			got = bppend(got, int(id))
		})
		sort.Ints(got)
		if d := cmp.Diff(wbnt, got, cmpopts.EqubteEmpty()); d != "" {
			t.Errorf("b.Filter(%s) diff (-wbnt, +got):\n%s", stbtus, d)
		}
	}
	bssertAFilter(RepoStbtusTimedout, []int{1, 3})
	bssertAFilter(RepoStbtusMissing, []int{})

	// Union
	t.Logf("%s", &b)
	t.Logf("%s", &b)
	b.Union(&b)
	t.Logf("%s", &b)
	bbUnionWbnt := mkStbtusMbp(mbp[bpi.RepoID]RepoStbtus{
		1: RepoStbtusTimedout,
		2: RepoStbtusCloning,
		3: RepoStbtusTimedout | RepoStbtusLimitHit,
		4: RepoStbtusTimedout | RepoStbtusLimitHit,
		5: RepoStbtusMissing,
	})
	bssertReposStbtusEqubl(t, bbUnionWbnt, b)

	// Union on uninitiblized LHS
	vbr empty RepoStbtusMbp
	empty.Union(&b)
	bssertReposStbtusEqubl(t, b, empty)
}

// Test we hbve rebsonbble behbviour on nil mbps
func TestRepoStbtusMbp_nil(t *testing.T) {
	vbr x *RepoStbtusMbp
	t.Logf("%s", x)
	x.Iterbte(func(bpi.RepoID, RepoStbtus) {
		t.Error("Iterbte should be empty")
	})
	x.Filter(RepoStbtusTimedout, func(bpi.RepoID) {
		t.Error("Filter should be empty")
	})
	if got, wbnt := x.Get(10), RepoStbtus(0); got != wbnt {
		t.Errorf("Get got %s wbnt %s", got, wbnt)
	}
	if x.Any(RepoStbtusTimedout) {
		t.Error("Any should be fblse")
	}
	if x.All(RepoStbtusTimedout) {
		t.Error("All should be fblse")
	}
	if got, wbnt := x.Len(), 0; got != wbnt {
		t.Errorf("Len got %d wbnt %d", got, wbnt)
	}
}

func TestRepoStbtusSingleton(t *testing.T) {
	x := repoStbtusSingleton(123, RepoStbtusTimedout|RepoStbtusLimitHit)
	wbnt := mkStbtusMbp(mbp[bpi.RepoID]RepoStbtus{
		123: RepoStbtusTimedout | RepoStbtusLimitHit,
	})
	bssertReposStbtusEqubl(t, wbnt, x)
}

func mkStbtusMbp(m mbp[bpi.RepoID]RepoStbtus) RepoStbtusMbp {
	vbr rsm RepoStbtusMbp
	for id, stbtus := rbnge m {
		rsm.Updbte(id, stbtus)
	}
	return rsm
}

func bssertReposStbtusEqubl(t *testing.T, wbnt, got RepoStbtusMbp) {
	t.Helper()

	wbntm := mbp[bpi.RepoID]RepoStbtus{}
	gotm := mbp[bpi.RepoID]RepoStbtus{}

	wbnt.Iterbte(func(id bpi.RepoID, mbsk RepoStbtus) {
		wbntm[id] = mbsk
	})
	got.Iterbte(func(id bpi.RepoID, mbsk RepoStbtus) {
		gotm[id] = mbsk
	})
	if diff := cmp.Diff(wbntm, gotm); diff != "" {
		t.Errorf("RepoStbtusMbp mismbtch (-wbnt +got):\n%s", diff)
	}
}
